Add 'spectrum' visualizer.
The 'spectrum' program generates MIDI or Alias/Wavefront mesh files from a sound input, displaying frequency amplitudes over time.  It utilizes per-note digital filter convolution.

+++ src-library/Binasc.cpp
@@ -1374,9 +1374,9 @@ int Binasc::processDecimalWord(ostream& out, const string& word, int lineNum) {
             return 0;
          }
          ulong tempLong = (ulong)atoi(&word[quoteIndex + 1]);
-         uchar byte1 = (tempLong & 0x00ff0000) >> 16;
-         uchar byte2 = (tempLong & 0x0000ff00) >>  8;
-         uchar byte3 = (tempLong & 0x000000ff);
+         uchar byte1 = (uchar)((tempLong & 0x00ff0000) >> 16);
+         uchar byte2 = (uchar)((tempLong & 0x0000ff00) >>  8);
+         uchar byte3 = (uchar)((tempLong & 0x000000ff));
          if (endianIndex == -1) {
             out << byte1;
             out << byte2;
